Separated view tests to view specific modules
